lint fix

+++ vnext/Microsoft.ReactNative/Fabric/Composition/ScrollViewComponentView.h
@@ -147,7 +147,7 @@ struct ScrollInteractionTrackerOwner : public winrt::implements<
   bool m_isHorizontal = false;
   bool m_changeViewAfterLoaded = false;
   bool m_dismissKeyboardOnDrag = false;
-  float m_scrollEventThrottle{0.0f};
+  double m_scrollEventThrottle{0.0};
   bool m_allowNextScrollNoMatterWhat{false};
   std::chrono::steady_clock::time_point m_lastScrollEventTime{};
   std::shared_ptr<facebook::react::ScrollViewShadowNode::ConcreteState const> m_state;
